published scheme Client interface and helpers
